Add home fragment
