+++ app/src/main/java/me/tylerbwong/stack/ui/MainViewModel.kt
@@ -8,23 +8,10 @@ import me.tylerbwong.stack.data.auth.AuthRepository
 import me.tylerbwong.stack.data.auth.AuthStore
 import me.tylerbwong.stack.data.auth.LogOutResult.LogOutError
 import me.tylerbwong.stack.data.auth.LogOutResult.LogOutSuccess
-import me.tylerbwong.stack.data.model.CREATION
-import me.tylerbwong.stack.data.model.Question
-import me.tylerbwong.stack.data.model.Sort
-import me.tylerbwong.stack.data.network.ServiceProvider
-import me.tylerbwong.stack.data.network.service.QuestionService
-import me.tylerbwong.stack.data.repository.QuestionRepository
 import retrofit2.HttpException
 
-internal class MainViewModel(
-    private val authRepository: AuthRepository = AuthRepository(),
-    private val repository: QuestionRepository = QuestionRepository(),
-    private val service: QuestionService = ServiceProvider.questionService
-) : BaseViewModel() {
-
-    internal val questions: LiveData<List<Question>>
-        get() = _questions
-    private val _questions = MutableLiveData<List<Question>>()
+internal class MainViewModel(private val authRepository: AuthRepository = AuthRepository()) :
+    BaseViewModel() {
 
     internal val profileImage: LiveData<String?>
         get() = _profileImage
@@ -32,32 +19,6 @@ internal class MainViewModel(
 
     internal val isAuthenticated: LiveData<Boolean>
         get() = AuthStore.isAuthenticatedLiveData
-
-    @Sort
-    internal var currentSort: String = CREATION
-    internal var currentQuery: String = ""
-
-    internal fun getQuestions(@Sort sort: String = currentSort) {
-        currentSort = sort
-        launchRequest {
-            _questions.value = repository.getQuestions(sort)
-        }
-    }
-
-    internal fun searchQuestions(query: String = currentQuery) {
-        currentQuery = query
-        launchRequest {
-            _questions.value = service.getQuestionsBySearchString(searchString = query).items
-        }
-    }
-
-    internal fun fetchQuestions() {
-        if (currentQuery.isNotBlank()) {
-            searchQuestions()
-        } else {
-            getQuestions()
-        }
-    }
 
     internal fun fetchUser() {
         viewModelScope.launch {
@@ -77,19 +38,8 @@ internal class MainViewModel(
                 is LogOutSuccess -> {
                     _profileImage.value = null
                     mutableSnackbar.value = null
-                    fetchQuestions()
                 }
             }
         }
     }
-
-    internal fun onQueryTextChange(newText: String?) {
-        currentQuery = newText ?: ""
-
-        if (currentQuery.isBlank()) {
-            getQuestions()
-        }
-    }
-
-    internal fun isQueryBlank() = currentQuery.isBlank()
 }
