+++ app/src/main/res/layout/header_holder.xml
@@ -15,9 +15,9 @@
         style="@style/TextAppearance.MaterialComponents.Headline4"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/questions"
         android:textColor="?android:attr/textColorPrimary"
-        android:textStyle="bold" />
+        android:textStyle="bold"
+        tools:text="@string/questions" />
 
     <TextView
         android:id="@+id/subtitle"
